refactor: better class design
Move CPU-only code down to the PatchFactoryCpu class instead of the PatchFactoryBase class.
