+++ src/lib/terrain/patch/patch-factory/merged/gpu/patch-factory-gpu.ts
@@ -1,5 +1,4 @@
 import { type IVoxelMap } from '../../../../i-voxel-map';
-import { EPatchComputingMode } from '../../patch-factory-base';
 import { PatchFactory } from '../patch-factory';
 import { type PatchSize } from '../vertex-data1-encoder';
 
@@ -8,11 +7,8 @@ import { PatchComputerGpu } from './patch-computer-gpu';
 abstract class PatchFactoryGpu extends PatchFactory {
     private readonly patchComputerGpuPromise: Promise<PatchComputerGpu> | null = null;
 
-    protected constructor(map: IVoxelMap, computingMode: EPatchComputingMode, patchSize: PatchSize) {
-        if (computingMode !== EPatchComputingMode.GPU_SEQUENTIAL && computingMode !== EPatchComputingMode.GPU_OPTIMIZED) {
-            throw new Error(`Unsupported computing mode "${computingMode}".`);
-        }
-        super(map, computingMode, patchSize);
+    public constructor(map: IVoxelMap, patchSize: PatchSize) {
+        super(map, patchSize);
         const localCacheSize = this.maxPatchSize.clone().addScalar(2);
         this.patchComputerGpuPromise = PatchComputerGpu.create(localCacheSize, this.vertexData1Encoder, PatchFactory.vertexData2Encoder);
     }
