+++ src/lib/terrain/patch/patch-factory/patch-factory-base.ts
@@ -5,7 +5,6 @@ import type { PatchMaterialUniforms, PatchMaterials } from '../material';
 import { Patch } from '../patch';
 import { PatchId } from '../patch-id';
 
-import * as Cube from './cube';
 import type { PackedUintFragment } from './uint-packing';
 
 type GeometryAndMaterial = {
@@ -23,36 +22,16 @@ type VertexData = {
     readonly roundnessY: boolean;
 };
 
-type FaceData = {
-    readonly voxelLocalPosition: THREE.Vector3;
-    readonly voxelMaterialId: number;
-    readonly faceType: Cube.FaceType;
-    readonly faceId: number;
-    readonly verticesData: [VertexData, VertexData, VertexData, VertexData];
-};
-
 type LocalMapData = {
+    readonly size: THREE.Vector3;
     readonly data: Uint16Array;
     readonly isEmpty: boolean;
 };
-
-type LocalMapCache = LocalMapData & {
-    readonly size: THREE.Vector3;
-    neighbourExists(voxelIndex: number, neighbourRelativePosition: THREE.Vector3): boolean;
-};
-
-enum EPatchComputingMode {
-    CPU_CACHED,
-    GPU_SEQUENTIAL,
-    GPU_OPTIMIZED,
-}
 
 abstract class PatchFactoryBase {
     public static readonly maxSmoothEdgeRadius = 0.3;
 
     public abstract readonly maxPatchSize: THREE.Vector3;
-
-    private readonly computingMode: EPatchComputingMode;
 
     protected readonly map: IVoxelMap;
 
@@ -64,9 +43,8 @@ abstract class PatchFactoryBase {
 
     protected readonly uniformsTemplate: PatchMaterialUniforms;
 
-    protected constructor(map: IVoxelMap, voxelTypeEncoder: PackedUintFragment, computingMode: EPatchComputingMode) {
+    protected constructor(map: IVoxelMap, voxelTypeEncoder: PackedUintFragment) {
         this.map = map;
-        this.computingMode = computingMode;
 
         this.texture = PatchFactoryBase.buildMaterialsTexture(map.voxelMaterialsList, voxelTypeEncoder);
         this.noiseTexture = PatchFactoryBase.buildNoiseTexture(this.noiseResolution, this.noiseTypes);
@@ -154,102 +132,11 @@ abstract class PatchFactoryBase {
         );
     }
 
-    protected async iterateOnVisibleFaces(patchStart: THREE.Vector3, patchEnd: THREE.Vector3): Promise<() => Generator<FaceData>> {
-        const that = this;
-
-        if (this.computingMode === EPatchComputingMode.CPU_CACHED) {
-            const localMapCache = await this.buildLocalMapCache(patchStart, patchEnd);
-            return function* () {
-                for (const a of that.iterateOnVisibleFacesWithCache(localMapCache)) {
-                    yield a;
-                }
-            };
-        } else {
-            throw new Error(`Unsupported patch computing mode ${this.computingMode}`);
-        }
-    }
-
-    private *iterateOnVisibleFacesWithCache(localMapCache: LocalMapCache): Generator<FaceData> {
-        if (localMapCache.isEmpty) {
-            return;
-        }
-
-        let cacheIndex = 0;
-        const localPosition = new THREE.Vector3();
-        for (localPosition.z = 0; localPosition.z < localMapCache.size.z; localPosition.z++) {
-            for (localPosition.y = 0; localPosition.y < localMapCache.size.y; localPosition.y++) {
-                for (localPosition.x = 0; localPosition.x < localMapCache.size.x; localPosition.x++) {
-                    const cacheData = localMapCache.data[cacheIndex];
-                    if (typeof cacheData === 'undefined') {
-                        throw new Error();
-                    }
-
-                    if (cacheData > 0) {
-                        // if there is a voxel there
-                        if (
-                            localPosition.x > 0 &&
-                            localPosition.y > 0 &&
-                            localPosition.z > 0 &&
-                            localPosition.x < localMapCache.size.x - 1 &&
-                            localPosition.y < localMapCache.size.y - 1 &&
-                            localPosition.z < localMapCache.size.z - 1
-                        ) {
-                            const voxelLocalPosition = localPosition.clone().subScalar(1);
-                            const voxelMaterialId = cacheData - 1;
-
-                            for (const face of Object.values(Cube.faces)) {
-                                if (localMapCache.neighbourExists(cacheIndex, face.normal.vec)) {
-                                    // this face will be hidden -> skip it
-                                    continue;
-                                }
-
-                                yield {
-                                    voxelLocalPosition,
-                                    voxelMaterialId,
-                                    faceType: face.type,
-                                    faceId: face.id,
-                                    verticesData: face.vertices.map((faceVertex: Cube.FaceVertex): VertexData => {
-                                        let ao = 0;
-                                        const [a, b, c] = faceVertex.shadowingNeighbourVoxels.map(neighbourVoxel =>
-                                            localMapCache.neighbourExists(cacheIndex, neighbourVoxel)
-                                        ) as [boolean, boolean, boolean];
-                                        if (a && b) {
-                                            ao = 3;
-                                        } else {
-                                            ao = +a + +b + +c;
-                                        }
-
-                                        let roundnessX = true;
-                                        let roundnessY = true;
-                                        for (const neighbourVoxel of faceVertex.edgeNeighbourVoxels.x) {
-                                            roundnessX &&= !localMapCache.neighbourExists(cacheIndex, neighbourVoxel);
-                                        }
-                                        for (const neighbourVoxel of faceVertex.edgeNeighbourVoxels.y) {
-                                            roundnessY &&= !localMapCache.neighbourExists(cacheIndex, neighbourVoxel);
-                                        }
-
-                                        return {
-                                            localPosition: faceVertex.vertex,
-                                            ao,
-                                            roundnessX,
-                                            roundnessY,
-                                        };
-                                    }) as [VertexData, VertexData, VertexData, VertexData],
-                                };
-                            }
-                        }
-                    }
-                    cacheIndex++;
-                }
-            }
-        }
-    }
-
     protected abstract computePatchData(patchStart: THREE.Vector3, patchEnd: THREE.Vector3): Promise<GeometryAndMaterial[]>;
 
     protected abstract disposeInternal(): void;
 
-    protected async buildLocalMapCache(patchStart: THREE.Vector3, patchEnd: THREE.Vector3): Promise<LocalMapCache> {
+    protected async buildLocalMapData(patchStart: THREE.Vector3, patchEnd: THREE.Vector3): Promise<LocalMapData> {
         const cacheStart = patchStart.clone().subScalar(1);
         const cacheEnd = patchEnd.clone().addScalar(1);
         const cacheSize = new THREE.Vector3().subVectors(cacheEnd, cacheStart);
@@ -261,25 +148,8 @@ abstract class PatchFactoryBase {
             throw new Error(`Invalid cache length. Should be ${expectedCacheItemsCount} items but is ${localMapData.data.length} items`);
         }
 
-        const indexFactor = { x: 1, y: cacheSize.x, z: cacheSize.x * cacheSize.y };
-
-        const buildIndexUnsafe = (position: THREE.Vector3) => {
-            return position.x * indexFactor.x + position.y * indexFactor.y + position.z * indexFactor.z;
-        };
-
-        const neighbourExists = (index: number, neighbour: THREE.Vector3) => {
-            const deltaIndex = buildIndexUnsafe(neighbour);
-            const neighbourIndex = index + deltaIndex;
-            const neighbourData = localMapData.data[neighbourIndex];
-            if (typeof neighbourData === 'undefined') {
-                throw new Error();
-            }
-            return neighbourData !== 0;
-        };
-
         return Object.assign(localMapData, {
             size: cacheSize,
-            neighbourExists,
         });
     }
 
@@ -324,4 +194,4 @@ abstract class PatchFactoryBase {
     }
 }
 
-export { EPatchComputingMode, PatchFactoryBase, type GeometryAndMaterial, type LocalMapCache, type VertexData };
+export { PatchFactoryBase, type GeometryAndMaterial, type LocalMapData, type VertexData };
